minor clean
